requirements.txt
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
streamlit
pillow
plotly
joblib
